changed the cart component structure
